+++ src/classes/BigCard.js
