+++ src/img/loading.gif
